[CLI] Add more flexibility to what can be watched
If no paths are defined, then watch `lotus.path`.

If one path is defined, resolve it and look for a "package.json" file.

If the path is not a directory, throw an error.

If the directory has a "package.json" file, treat it like a module.

Otherwise, treat the path like a directory of modules.